+++ src/client/redux/actionTypes.ts
@@ -12,3 +12,4 @@ export const SET_TOTAL_PAGES = "SET_TOTAL_PAGES";
 // * User
 export const SET_EMAIL = "SET_EMAIL";
 export const SET_IS_LOGGED_IN = "SET_IS_LOGGED_IN";
+export const SET_PASSWORD = "SET_PASSWORD";