+++ src/client/redux/actions/user.ts
@@ -1,4 +1,4 @@
-import { SET_EMAIL, SET_IS_LOGGED_IN } from "../actionTypes";
+import { SET_EMAIL, SET_IS_LOGGED_IN, SET_PASSWORD } from "../actionTypes";
 
 import { UserAction } from "../../types";
 
@@ -10,4 +10,9 @@ export const setEmail = (email: string): UserAction => ({
 export const setIsLoggedIn = (isLoggedIn: boolean): UserAction => ({
   type: SET_IS_LOGGED_IN,
   payload: { isLoggedIn },
+});
+
+export const setPassword = (password: string): UserAction => ({
+  type: SET_PASSWORD,
+  payload: { password },
 });